Add files via upload
This is the CSV file that the ProducListCSV class should read from. Store it somewhere and copy the pathway to this file and then paste it in the ProductListCSV class constructor in replace of the other path. 
i.e.   this.file = "C:\\Users\\Gorilla Rig\\OneDrive - AUT University\\Documents\\NetBeansProjects\\InventoryManagementSystems\\resources\\ProductListFile.csv";
replace the assignment with your pathway